feat : 오브젝트 chapter10 통화요금 세금 정책 요구사항 추가 구현

+++ book-objects/src/main/java/chapter10/Money.java
@@ -14,7 +14,7 @@ public class Money {
         return new Money(BigDecimal.valueOf(amount));
     }
 
-    public Money times(long percent) {
+    public Money times(double percent) {
         return new Money(amount.multiply(BigDecimal.valueOf(percent)));
     }
